Cache: Remove SetSourceModule override, add ResetCache
Resetting the cache by overriding SetSourceModule doesn't work very well. The source modules of the Cache's source modules might be changed,
or the properties of any of the Cache's source modules might be changed, and none of these cases are handled.

So instead add a ResetCache method and require the user to reset the cache manually when needed.